recipe quantities and units are now float type

+++ src/Entity/RecipeFood.php
@@ -18,7 +18,7 @@ class RecipeFood
     private $id;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     public $quantity;
 
@@ -60,12 +60,12 @@ class RecipeFood
         return $this->id;
     }
 
-    public function getQuantity(): ?int
+    public function getQuantity(): ?float
     {
         return $this->quantity;
     }
 
-    public function setQuantity(int $quantity): self
+    public function setQuantity(float $quantity): self
     {
         $this->quantity = $quantity;
